refactor: extract MaintenanceEditModal
